Move GetBrowserLanguage service to ngeo/utils
