+++ contribs/gmf/test/.eslintrc.yaml
@@ -1,0 +1,2 @@
+env:
+  jasmine: true
