+++ package.json
@@ -53,6 +53,7 @@
     "@types/jqueryui": "^1.12.7",
     "@types/jsts": "^0.17.0",
     "@types/proj4": "^2.5.0",
+    "@types/sinon": "^7.0.13",
     "@types/typeahead": "^0.11.32",
     "angular": "1.7.8",
     "angular-animate": "^1.7.8",
@@ -98,6 +99,7 @@
     "karma": "^4.1.0",
     "karma-chrome-launcher": "^2.2.0",
     "karma-jasmine": "^2.0.1",
+    "karma-sinon": "^1.0.5",
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^4.0.2",
     "loader-utils": "^1.2.3",
@@ -115,6 +117,7 @@
     "proj4": "^2.5.0",
     "raven-js": "^3.27.1",
     "simple-html-tokenizer": "^0.1.1",
+    "sinon": "^7.3.2",
     "svg-inline-loader": "^0.8.0",
     "svgo": "^1.2.2",
     "svgo-loader": "^2.2.0",
