+++ test/.eslintrc.yaml
@@ -1,0 +1,4 @@
+env:
+  jasmine: true
+globals:
+  sinon: true
