+++ test/spec/all.js
@@ -2,11 +2,11 @@ import 'jquery';
 import 'angular';
 import 'angular-mocks';
 import 'angular-gettext';
+import './ngeo/getbrowserlanguage.spec.js';
 import './beforeeach.js';
 import './search/createlocationsearchbloodhound.spec.js';
 import './services/wmstime.spec.js';
 import './services/layerHelper.spec.js';
-import './services/getbrowserlanguage.spec.js';
 import './services/backgroundlayermgr.spec.js';
 import './services/decoratelayer.spec.js';
 import './services/wfspermalink.spec.js';
